RedrawCanvas.ts: reduce allocations
Perf measurements showed that a lot of time was spent in GC. Reducing
the amount of garbage generated makes everything snappy and smooth.
The main places where we allocate memory now is in the messagepack
encoding/decoding functions in Stdout.ts. There isn't much of point in
trying to optimize that as the bottleneck shifted back to the rendering
routines.
The bottleneck now is the use of putImageData, which is faster than
manually calling fillRect+fillText but much slower than what could be
achieved with webgl. But performance is ok and putImageData is much
simpler than webgl, so there's so the renderer is in an ok state in my
opinion.